Implement CPU-GPU micro-mastery remediation plan, Phases 0-3, 5, 4.1-4.2
Implements docs/CPU_GPU_MICRO_MASTERY_REMEDIATION_PLAN.md (Claude Opus 5's
review of 9a9e9d85/a291cb10/8bd89909), leaving Phase 4.3 (instancing) for
Zach per the plan's own AUTHOR flag.

Phase 0: webgpu_micro_mastery_lag_test gets RUN_SERIAL/LABELS "lag" like
frame_lag_test, plus a calibration-normalised baseline
(tests/singularity/webgpu_micro_mastery_lag_baseline.txt).

Phase 1: WebGpuRenderer::bindPipeline skips redundant SetPipeline calls.
pipelineSwitches: ~19,500/frame -> 2.

Phase 2: TessMesh gains a `revision` counter checked by GpuMeshCache
alongside `id` (guards a future in-place mutator; no current writer needs
it since every writer today replaces the whole mesh). GpuBufferPool::init()
calls shutdown() first so a second init can't leak the previous device's
chunks.

Phase 3: gpu_mastery_test gains a device-backed section verifying both
audit-claimed fixes, frame-based GC, and the revision guard by reading
buffer CONTENTS back (WGPUBufferUsage_CopySrc added to the mesh cache's
vertex buffers) rather than comparing WGPUBuffer pointers, which the driver
can recycle after a release.

Phase 4.1-4.2: Object::drawCube merges all six faces into one draw when
they resolve identical paint (the unpainted case), decided fresh every
draw so a single setFaceColor stroke drops back to six draws. Guarded in
webgpu_object_test.

Phase 5: the seven derived @screen-channel metrics become read-only
ComputedProperty (wireframe stays writable); VRAM/uniform byte counts
become double instead of float.

Corrected two overstated claims in CPU_GPU_MICRO_MASTERY.md (the linear
search doesn't saturate intra-frame; the pool's safety is a WebGPU queue-
ordering property, not a design property).

Measured (webgpu_micro_mastery_lag_test, 4,500 objects): draw calls
19,500 -> 4,500/frame, pipeline switches ~19,500 -> 2/frame, average frame
time ~175ms -> ~104ms. Suite: 70/70.

+++ src/Singularity/Core/EngineRender.cpp
@@ -136,8 +136,8 @@ namespace Core {
                 const auto& stats = currentRenderer().frameStats();
                 sc->updateMetrics(static_cast<int>(stats.drawCalls),
                                   static_cast<int>(stats.trianglesDrawn),
-                                  static_cast<float>(stats.vramAllocatedBytes),
-                                  static_cast<float>(stats.uniformBytesWritten),
+                                  static_cast<double>(stats.vramAllocatedBytes),
+                                  static_cast<double>(stats.uniformBytesWritten),
                                   static_cast<int>(stats.bufferSuballocations),
                                   static_cast<int>(stats.pipelineSwitches),
                                   static_cast<int>(stats.cachedMeshesCount));
